+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 1.0.1-202201xx
+* Packer: add cleanup 
+* Ansible: 5.2.0
+* VirtualBox guest additions: 6.1.30
+
 ### 1.0.0-20220128
 * First release of this box, based on Ubuntu 20.04.3
 * Ansible: 5.2.0